requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
aiosqlite
pydantic[email]
pydantic-settings
python-dotenv
python-multipart
python-jose[cryptography]
bcrypt
